@@ -4,9 +4,7 @@ import { useEffect } from 'react';
 import s from 'styles/pages/404.module.css';
 
 export default function Eror({ error, reset }) {
-  useEffect(() => {
-    console.error(error);
-  }, [error]);
+  useEffect(() => {}, [error]);
 
   return (
     <>
