@@ -45,7 +45,7 @@ export default async function OpenSourcePage() {
       <Container clean name="you">
         <div className={s.thanksContainer}>
           <Fade direction="left" clean className="w-full">
-            <h2 className="text-4xl font-bold text-center">Thank you.</h2>
+            <h2 className="text-center font-bold text-4xl">Thank you.</h2>
 
             <Sponsors sponsors={sponsors} />
           </Fade>
@@ -56,7 +56,7 @@ export default async function OpenSourcePage() {
       <Container clean name="people">
         <div className={s.genericContainer}>
           <Fade direction="left" delay={0.5} className="w-full">
-            <h2 className="text-4xl font-bold text-center">Still You</h2>
+            <h2 className="text-center font-bold text-4xl">Still You</h2>
 
             <Followers followers={followers} />
           </Fade>
@@ -67,7 +67,7 @@ export default async function OpenSourcePage() {
       <Container clean name="passion">
         <div className={s.genericContainer}>
           <Fade direction="left" delay={0.5} className="w-full">
-            <h2 className="text-4xl font-bold text-center">About me</h2>
+            <h2 className="text-center font-bold text-4xl">About me</h2>
 
             <GitHubProfile {...profile} />
           </Fade>
