@@ -1,6 +1,6 @@
 export default function Footer() {
   return (
-    <footer className="p-4 mt-auto text-right text-white sm:text-center">
+    <footer className="mt-auto p-4 text-right text-white sm:text-center">
       <p className="absolute left-0 ml-2 text-xs">
         Made with{' '}
         <span role="img" aria-label="heart">
