@@ -4,7 +4,7 @@ export default function ChevronUp({ ...props }) {
       <svg
         alt="up"
         xmlns="http://www.w3.org/2000/svg"
-        className="w-6 h-6"
+        className="h-6 w-6"
         fill="none"
         viewBox="0 0 24 24"
         stroke="currentColor"
