@@ -5,7 +5,7 @@ export default function Fork({ ...props }) {
         <svg
           alt="fork"
           xmlns="http://www.w3.org/2000/svg"
-          className="w-6 h-6"
+          className="h-6 w-6"
           fill="none"
           viewBox="0 0 16 16"
           stroke="currentColor"
