@@ -6,7 +6,7 @@ export default function Github({ ...props }) {
         fill="#fff"
         xmlns="http://www.w3.org/2000/svg"
         viewBox="0 0 24 24"
-        className="w-6 h-6"
+        className="h-6 w-6"
         {...props}
       >
         <title>GitHub</title>
