@@ -36,12 +36,12 @@ const Player = () => {
             href={url}
           >
             {listening?.isPlaying ? (
-              <div className="w-auto h-auto">
+              <div className="h-auto w-auto">
                 {/* eslint-disable-next-line @next/next/no-img-element */}
                 <img width="40" height="40" src={listening?.thumbnail} alt={listening?.album} />
               </div>
             ) : (
-              <Spotify className="w-10 h-10" color={'#1ED760'} />
+              <Spotify className="h-10 w-10" color={'#1ED760'} />
             )}
           </Link>
 
@@ -60,7 +60,7 @@ const Player = () => {
                   rel="noopener noreferer noreferrer"
                   title="Mateo Nunez on Spotify"
                 >
-                  <ChevronUp className="w-4 h-4 rotate-90" />
+                  <ChevronUp className="h-4 w-4 rotate-90" />
                 </Link>
               </div>
             </div>
