@@ -1,5 +1,5 @@
-import { promises } from 'fs';
-import path from 'path';
+import { promises } from 'node:fs';
+import path from 'node:path';
 
 import matter from 'gray-matter';
 import removeMarkdown from 'remove-markdown';
@@ -90,7 +90,7 @@ const serializeOptions = {
     remarkPlugins,
     rehypePlugins,
   },
-}
+};
 
 async function compileSource({ content }) {
   const { compiledSource } = await serialize(content, serializeOptions);
