@@ -17,7 +17,19 @@ const metadata = {
     default: defaultTitle,
   },
   description: defaultDescription,
-  keywords: ['mateo nunez', 'mateonunez', 'software engineer', 'open source', 'javascript', 'nodejs', 'react', 'nextjs', 'typescript', 'colombia', 'italy'],
+  keywords: [
+    'mateo nunez',
+    'mateonunez',
+    'software engineer',
+    'open source',
+    'javascript',
+    'nodejs',
+    'react',
+    'nextjs',
+    'typescript',
+    'colombia',
+    'italy',
+  ],
   author,
   authors: [author],
   colorSchema: 'dark',
