@@ -4,9 +4,8 @@ import { useEffect } from 'react';
 
 export const useTimeline = (options) => gsap.timeline(options);
 
-// biome-ignore lint/nursery/noEmptyBlockStatements: <explanation>
 export const useAnimation = (ref, animation = () => {}, timelines = []) => {
-  const timeline = !timelines.length ? gsap.timeline() : timelines[0];
+  const timeline = timelines.length === 0 ? gsap.timeline() : timelines[0];
 
   const { current } = ref;
   const { from, to } = animation();
